Cache types use traits for test cases
All test cases have been moved to a single trait for convinence in
future implementations.
