+++ src/ArrayCache.php
@@ -33,6 +33,10 @@ class ArrayCache implements CacheInterface
         $this->validateKey($key);
         $this->data[$key] = $value;
 
+        if (is_null($value)) {
+            return false;
+        }
+
         return true;
     }
 
@@ -78,32 +82,5 @@ class ArrayCache implements CacheInterface
         $this->validateKey($key);
 
         return isset($this->data[$key]);
-    }
-
-    private function transformKeys($keys)
-    {
-        if ($keys instanceof Traversable) {
-            return $this->transformTraversableKeys($keys);
-        } elseif (is_array($keys)) {
-            foreach ($keys as $key) {
-                $this->validateKey($key);
-            }
-            return $keys;
-        }
-
-        throw new InvalidArgumentException(
-            "Cannot call getMultiple with a non array or non Traversable type"
-        );
-    }
-
-    private function transformTraversableKeys($keys)
-    {
-        $tempKeys = [];
-        foreach ($keys as $key => $value) {
-            $this->validateKey($key);
-            $tempKeys[] = $key;
-        }
-
-        return $tempKeys;
     }
 }
