+++ tests/CacheInterfaceTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Realpage\SimpleCache\Tests;
+
+use Psr\SimpleCache\CacheInterface;
+
+interface CacheInterfaceTest
+{
+    public function buildCache($data = null): CacheInterface;
+}
